backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0
pymysql==1.1.0
python-dotenv==1.0.0
Werkzeug==3.0.1
Pillow==10.1.0
cryptography==41.0.7
